+++ src/test/java/com/epam/reportportal/testng/ParametersBypassTest.java
@@ -25,6 +25,7 @@ import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 import static com.epam.reportportal.testng.integration.util.TestUtils.namedUuid;
+import static com.epam.reportportal.testng.integration.util.TestUtils.standardParameters;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.*;
 import static org.mockito.ArgumentMatchers.startsWith;
@@ -72,7 +73,7 @@ public class ParametersBypassTest {
 	public void initMocks() {
 		TestUtils.mockLaunch(client, "launchUuid", suitedUuid, testClassUuid, testMethodUuidList);
 
-		final ReportPortal reportPortal = ReportPortal.create(client, new ListenerParameters(PropertiesLoader.load()));
+		final ReportPortal reportPortal = ReportPortal.create(client, standardParameters());
 		TestReportPortalListener.initReportPortal(reportPortal);
 	}