+++ src/test/java/com/epam/reportportal/testng/TestLaunchFinishShutdownHook.java
@@ -1,5 +1,6 @@
 package com.epam.reportportal.testng;
 
+import com.epam.reportportal.service.statistics.StatisticsService;
 import com.epam.reportportal.testng.integration.feature.shutdown.LaunchFinishShutdownHookRemoveTest;
 import com.epam.reportportal.testng.integration.feature.shutdown.LaunchFinishShutdownHookTest;
 import com.epam.reportportal.util.test.ProcessUtils;
@@ -26,14 +27,15 @@ public class TestLaunchFinishShutdownHook {
 
 		ServerSocket ss = SocketUtils.getServerSocketOnFreePort();
 		SocketUtils.ServerCallable serverCallable = new SocketUtils.ServerCallable(ss, Collections.emptyMap(), "files/socket_response.txt");
-		Callable<Process> clientCallable = () -> ProcessUtils.buildProcess(true, clazz, String.valueOf(ss.getLocalPort()));
+		Callable<Process> clientCallable = () -> ProcessUtils.buildProcess(true, clazz,
+				Collections.singletonMap(StatisticsService.DISABLE_PROPERTY, "1"), String.valueOf(ss.getLocalPort()));
 		Pair<String, Process> startResult = SocketUtils.executeServerCallable(serverCallable, clientCallable);
 		assertThat(startResult.getValue(), notNullValue());
 		assertThat("First request is a launch start", startResult.getKey(), startsWith("POST /api/v1/test-project/launch"));
 
 		Callable<Integer> clientCallableResult = () -> {
 			try {
-				if (startResult.getValue().waitFor(6, TimeUnit.SECONDS)) {
+				if (startResult.getValue().waitFor(7, TimeUnit.SECONDS)) {
 					return startResult.getValue().exitValue();
 				} else {
 					startResult.getValue().destroy();